requirements.txt
fastapi==0.127.0
uvicorn[standard]==0.40.0
httpx==0.27.2
python-dotenv==1.2.1
pydantic==2.12.5

